Release 2.3.2

+++ buession-dao/src/main/java/com/buession/dao/mybatis/dialect/Dialect.java
@@ -60,7 +60,7 @@ public interface Dialect {
 	 *
 	 * @return 分页 SQL
 	 */
-	default String buildPaginationSql(final String sql, final long offset, final long limit) {
+	default String buildPaginationSql(final String sql, final int offset, final int limit) {
 		return null;
 	}
